Move StdOutWritter in new pacakge logging.
