+++ jflex/src/main/java/jflex/core/BUILD
@@ -13,6 +13,7 @@ java_library(
         "//jflex/src/main/java/jflex/exceptions",
         "//jflex/src/main/java/jflex/io",
         "//jflex/src/main/java/jflex/l10n",
+        "//jflex/src/main/java/jflex/logging",
         "//jflex/src/main/java/jflex/performance",
         "//jflex/src/main/java/jflex/scanner",
     ],
