+++ jflex/src/main/java/jflex/core/Out.java
@@ -19,6 +19,7 @@ import jflex.base.Build;
 import jflex.core.unicode.UnicodeProperties;
 import jflex.exceptions.GeneratorException;
 import jflex.l10n.ErrorMessages;
+import jflex.logging.StdOutWriter;
 import jflex.performance.Timer;
 
 /**
