+++ jflex/src/main/java/jflex/logging/BUILD
@@ -1,0 +1,6 @@
+java_library(
+    name = "logging",
+    srcs = ["StdOutWriter.java"],
+    visibility = ["//jflex:__subpackages__"],
+    deps = [],
+)
